+++ src/design.txt
@@ -13,7 +13,7 @@ is manageable by one developer.
 GTT four basic parts:
 1) the 'engine', a set of C objects that encapsulate 
    project, task and time interval data with a set of 
-   GUI-independent programming API's.  The API is in proj.h,
+   GUI-independent programming API's.  The API is in gtt_project.h,
    the code in proj.c and the private data structs in 
    proj_p.h.  The goal of splitting out the data this way is
    to make it easier to write an SQL back-end for storage,
