+++ src/gtt_activation_dialog.c
@@ -28,7 +28,7 @@
 
 #include "cur-proj.h"
 #include "gtt_help_popup.h"
-#include "proj.h"
+#include "gtt_project.h"
 #include "proj-query.h"
 #include "gtt_util.h"
