+++ src/gtt_journal.c
@@ -35,9 +35,9 @@
 #include "app.h"
 #include "ghtml.h"
 #include "gtt_help_popup.h"
-#include "menus.h"
+#include "gtt_menus.h"
 #include "plug-in.h"
-#include "proj.h"
+#include "gtt_project.h"
 #include "props-invl.h"
 #include "props-task.h"
 #include "gtt_util.h"
