+++ src/gtt_plug_in_edit.c
@@ -25,7 +25,7 @@
 #include "gtt_help_popup.h"
 #include "gtt_journal.h"
 #include "gtt_gsettings_io.h"
-#include "menus.h"
+#include "gtt_menus.h"
 #include "plug-in.h"
 #include "gtt_util.h"
