+++ src/gtt_props_dlg_task.c
@@ -23,7 +23,7 @@
 #include <string.h>
 
 #include "gtt_help_popup.h"
-#include "proj.h"
+#include "gtt_project.h"
 #include "props-task.h"
 #include "gtt_util.h"
